Fix MSVC14 warning C4800 bool conversion
This changes a boolean expression to not forcing an integer value into a
boolean value. The expression compares if two lengths are both zero or
both not zero. This had been done by casting the length into a boolean and
is changed by this commit to an explicit boolean check whether 'length > 0'.