Directly braze-initialize column constraints
Fixes #1196, a clang compiler bug in the mingw environment (tested with clang versions up to 16.0.2). Clang has troubles to construct a base's empty tuple member marked with the `[[no_unique_address]]` attribute from a temporary tuple. It works if the constraints tuple is also directly braze initialized (which is better coding anyway). Note that it would also work if we would remove the attribute, but that's not what we want.