No more including of `cxx_universal.h`
The `cxx_universal.h` exists to provide symbols that sqlite_orm depends on everywhere.
It is easy to forget to include it. And more importantly there are decision macros used that need to be defined before certain files anyway, which is the reason that `cxx_universal.h` gets included by the files used for amalgamation.
Hence I decided to not include it anymore.