SQLite error handling solely based on return code
The problem solved is the race condition between executing a statement and retrieving the last error from SQLite.
There are two ways to fix this: Either handle errors solely based on the return code of an API call. Or serialize access to the DB handle.

The core team of the library decided to implement handling errors based on the return code.

When handling errors solely based on the return code, SQLite obviously cannot provide as much information as to why (or where) a statement failed.

Important things to know:
* The library can handle empty ranges of insert/replace statements and unregistered function errors up-front.
* Checking missing application-defined collation sequences wasn't pursued because SQLite will at least return a dedicated `SQLITE_ERROR_MISSING_COLLSEQ` extended error value. It would be easily implementable.
* Detailed information of constraint errors [`SQLITE_CONSTRAINT_*], e.g. about the specific column violating a constraint (e.g. primary key, foreign key, not null) is lost; only the type of constraint violation is known.